.history/server/requirements_20240322152034.txt
quart==0.19.4
python-dotenv==1.0.1
quart_jwt_extended==0.1.0
pymongo==4.6.2
cerberus==1.3.5
bcrypt==4.1.2

langchain
sentence_transformers > /dev/null
langchain_experimental
pypdf
"unstructured[all-docs]"

#
# !apt-get install poppler-utils
# !apt-get update
# !apt-get install tesseract-ocr
#
rank_bm25 > /dev/null
faiss-gpu